requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
azure-ai-inference==1.0.0b9
azure-core==1.33.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-cosmos==4.9.0
azure-identity==1.21.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
dotenv==0.9.9
frozenlist==1.5.0
fsspec==2024.12.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
isodate==0.7.2
jiter==0.9.0
json_repair==0.39.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-azure-ai==0.1.2
langchain-core==0.3.51
langchain-ollama==0.3.0
langchain-openai==0.3.12
langchain-text-splitters==0.3.7
langsmith==0.3.18
msal==1.32.0
msal-extensions==1.3.1
multidict==6.3.2
neo4j==5.28.1
neo4j-graphrag==1.6.0
numpy==1.26.4
ollama==0.4.7
openai==1.71.0
opentelemetry-api==1.31.1
orjson==3.10.15
packaging==24.2
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11.3
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
shapely==2.1.0
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
urllib3==2.3.0
vertexai==1.71.1
wrapt==1.17.2
yarl==1.19.0
zipp==3.21.0
zstandard==0.23.0
